requirements/requirements.txt
# CORE
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0
python-dotenv>=0.15.0,<1.0.0
requests==2.26.0

